Tyuiu.PenkovEU.ConsoleApp.Lab0.V0.Lib/DataService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Tyuiu.PenkovEU.ConsoleApp.Lab0.V0.Lib
{
    public class DataService
    {
        public int Add(int x, int y)
        {
            return x + y;
        }

        public int Minus(int x, int y)
        {
            return x - y;
        }

        public int Multiplication(int x, int y)
        {
            return x * y;
        }

        public double Division(int x, int y)
        {
            return x / y;
        }
    }
}
